pyproject.toml
[project]
name = "etl-scripts"
# Semantic Versioning https://semver.org/
version = "0.1.0"
requires-python = ">=3.11"

# numcodecs, dask, pandas, numpy, and multiformats are imports in various files, but not in pyproject.toml dependencies, since they are required by xarray, so we just use the existing compatible versions that are installed when xarray is installed
dependencies = [
    "xarray[io,parallel]",
    "ipldstore @ git+https://github.com/dClimate/ipldstore@v2.1.3",
    "dc-etl @ git+https://github.com/dClimate/etl-tooling.git",
    "docopt",
    "copernicusmarine",
    "python-dotenv",
    "natsort",
    "deprecation",
    "fsspec",
    "kerchunk",
    "multiformats",
    "natsort",
    "nest-asyncio",
    "netcdf4<1.7.1",
    "numpy<2.0.0",
    "p_tqdm",
    "pandas",
    "psutil",
    "pycryptodome",
    "python-dateutil",
    "requests",
    "s3fs",
    "shapely",
    "statsmodels",
    "tqdm",
    "xarray[complete]>=2023.9.0",
    "zarr>=2.16.1",
    "cfgrib",
    "dag-cbor==0.2.2",
    "dask[array,diagnostics,distributed]",
    "lxml",
    "ipdb",
    "cdsapi==0.7.3",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit"
]
